docker-deployer/requirements.txt
bcrypt==3.2.0
boto3==1.17.44
botocore==1.20.112
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
Cython==0.29.22
docker==5.0.0
fast-json==0.3.2
idna==3.2
jmespath==0.10.0
mujson==1.4
paramiko==2.7.2
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
s3transfer==0.3.7
six==1.16.0
sshtunnel==0.4.0
toml==0.10.2
ujson==4.1.0
urllib3==1.26.6
websocket-client==1.2.1
